refactor(birdie): move birdieconfig
